kadai-tasklist3

+++ app/Http/Controllers/TasksController.php
@@ -15,12 +15,23 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //
-        $tasks = Task::all();
-        return view("tasks.index", [
-            "tasks" => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // task取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -49,11 +60,13 @@ class TasksController extends Controller
             "status" => "required|max:10",
             "content" => "required|max:255",
         ]);
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
         
+        $request->user()->tasks()->create([
+            "status" => $request->status,
+            'content' => $request->content,
+        ]);
+        
+        // 前のURLへリダイレクトさせる
         return redirect("/");
     }
 
@@ -66,11 +79,14 @@ class TasksController extends Controller
     public function show($id)
     {
         //
-        $task = Task::findOrFail($id);
+        $task = \App\Task::findOrFail($id);
         
-        return view("tasks.show", [
-            "task" => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view("tasks.show", [
+                "task" => $task,
+            ]);
+        }
+        return redirect("/");
     }
 
     /**
@@ -82,11 +98,15 @@ class TasksController extends Controller
     public function edit($id)
     {
         //
-        $task = Task::findOrFail($id);
+        $task = \App\Task::findOrFail($id);
         
-        return view("tasks.edit", [
-            "task" => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view("tasks.edit", [
+                "task" => $task,
+            ]);
+        }
+        return redirect("/");
+        
     }
 
     /**
@@ -103,10 +123,13 @@ class TasksController extends Controller
             "status" => "required|max:10",
             "content" => "required|max:255",
         ]);
-        $task = Task::findOrFail($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $task = \App\Task::findOrFail($id);
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+        }
         
         return redirect("/");
     }
@@ -120,9 +143,14 @@ class TasksController extends Controller
     public function destroy($id)
     {
         //
-        $task = Task::findOrFail($id);
-        $task->delete();
-        
+        $task = \App\Task::findOrFail($id);
+
+        // 認証済みユーザがその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
         return redirect("/");
     }
 }
